separated CLI functionality and logic
